% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rysuj.R
\name{rysuj}
\alias{rysuj}
\title{Funkcja tworzy wykres liczby komentarzy, udostępnień i polubień}
\usage{
rysuj(dane, ylab, title, crit)
}
\description{
Funkcja tworzy wykres liczby komentarzy, udostępnień i polubień
}

