% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wypiszDane.R
\name{wypiszDane}
\alias{wypiszDane}
\title{Funckja wypisuje ilosc komentarzy, udostepnien oraz ilosc polubień dla wybranego banku}
\usage{
wypiszDane(typ, bank, info)
}
\description{
Funckja wypisuje ilosc komentarzy, udostepnien oraz ilosc polubień dla wybranego banku
}

