% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wypiszPosty.R
\name{wypiszPosty}
\alias{wypiszPosty}
\title{Funkcja wypisuje wybraną ilość postów dla wybranego banku}
\usage{
wypiszPosty(bank, daty, ilosc, info)
}
\description{
Funkcja wypisuje wybraną ilość postów dla wybranego banku
}

